chore: 배너 공연목록 component로 변경
