+++ src/components/Card.jsx
@@ -1,10 +1,13 @@
+import { Link } from "react-router-dom";
 import "../components/styles/Card.css";
 
-function Card({ img_path, title, startDate, endDate, clubName }) {
+function Card({ id, img_path, title, startDate, endDate, clubName }) {
   return (
     <div id="list-card">
       <div id="list-image-box">
-        <img id="list-show-img" src={img_path} alt="show_image" />
+        <Link to={`/show/${id}`}>
+          <img id="list-show-img" src={img_path} alt="show_image" />
+        </Link>
       </div>
       <div id="list-info-box">
         <div id="list-title-box">{title} </div>
